+++ kount-access-java-sdk/src/test/java/com/kount/kountaccess/KountAccessExample.java
@@ -119,7 +119,7 @@ public class KountAccessExample {
 			String behavioHost = "api.behavio.kaptcha.com";
 			String environment = "sandbox";
 			String timing = "{\"valid\":\"json\"}";
-			sdk.setBehaviorData(behavioHost, environment, session, timing, uniq);
+			sdk.setBehavioData(behavioHost, environment, session, timing, uniq);
 
 			// info endpoint example (requesting all data)
 			int infoFlag = new InfoEndpointDataSet().withInfo().withVelocity().withBehavioSec().withDecision()