+++ RELEASE-NOTES.md
@@ -1,5 +1,26 @@
 kount-access-java-sdk 2.1.1
 ===========================
+**08/30/2018**
+
+### New features
+* default API version is changed from 2.1.0 to 4.0.0
+* added a few new endpoints
+** behavio data
+** device trust by device
+** device trust by session
+** get devices
+** get uniques
+** info
+
+### Improvements
+* updated examples
+* separated unit tests by endpoint
+* added more unit tests
+
+### Bugfixes
+
+kount-access-java-sdk 2.1.1
+===========================
 **07/14/2017**
 
 ### New features